Cast the searchField field to text in pgdefault.js to allow numeric fields
If the field specified by `searchField` in `models\pgdefault.js` refers to a numeric field, the SQL query fails with an error due to the `LOWER()` function not handling those. This casts the `searchField` field to TEXT in both the result set and the `WHERE` filter, to ensure that the lowercase functions works in both PostGIS and downstream in Origo.