Adds search option `useCentroid` to allow returning full feature geometry
Fixes #52.

Added a new option `useCentroid` (expects boolean) to `conf\dbconfig.js` search models that controls whether or not the geometries in the response will be converted to their centroids.

The geometry variables in the `models\...default.js` have been renamed from `centroid` to `wkt` to reflect this. In `models\oracledefault.js` the `wkt` variable was added so that the structure aligns with `mssqldefault.js` and `pgdefault.js`.

The `wkt` variable will default to using the centroid if the `useCentroid` option is missing.